Integrate demo.py and interfaces.py
