+++ MANIFEST.in
@@ -1,1 +1,1 @@
-include factoring/json_schema.json
+include json_schema.json
